Fix dependency resolution, support other scopes
Currently, we would only consider dependencies in the "compile" scope
for transpilation purposes.

This prevents building code that has annotations declared in
dependencies of "provided" or "system" scope, for example.

We would also erroneously include ".pom" files in the classpath, which
breaks the build in another place.

In addition to the "compile" scope, consider other scopes as well.
Furthermore, skip any entries in the classpath that end in ".pom".